Fix UI: dark mode, settings, button dedup, a11y, and engine extraction
Wire up theming and fix several dead or off-system UI controls, then
extract WordPathGame's pure engine layer into its own module.

Theming & settings
- Mount next-themes ThemeProvider (attribute="class") so the existing
  .dark palette and all dark: variants are actually reachable; toasts
  now follow the theme too.
- Replace the dead Theme and Font Size <select> controls in GameSettings
  with wired shadcn Select components. Font size persists via a new
  lib/appearance helper, applied before first paint in main.tsx.

Design system & a11y
- Add a `brand` button variant and replace ~13 copy-pasted brand class
  strings in WordPathGame with variant="brand".
- NotFound now uses design tokens instead of hardcoded gray/blue.
- Remove the two no-op Tutorial buttons from TitleScreen and add
  aria-label to the icon-only Settings buttons.
- TitleScreen uses min-h-[100dvh] + overflow-x-hidden so the button
  stack scrolls instead of clipping on short/landscape screens.

Refactor
- Extract the ~1035-line module-level engine (board generation, word
  probing, scoring, special-tile effects) from WordPathGame.tsx into
  wordPathEngine.ts. WordPathGame drops 5603 -> 4581 lines; the bundle
  chunk is unchanged and no runtime behavior changes.
