+++ src/App.tsx
@@ -3,6 +3,7 @@ import { Toaster as Sonner } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { ThemeProvider } from "next-themes";
 import { Suspense, lazy } from "react";
 import Index from "./pages/Index";
 
@@ -26,6 +27,7 @@ const LoadingSpinner = () => <div className="min-h-screen flex items-center just
   </div>;
 const queryClient = new QueryClient();
 const App = () => <QueryClientProvider client={queryClient}>
+    <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
     <SoundProvider>
       <ColorBlindProvider>
         <ARIAProvider>
@@ -55,5 +57,6 @@ const App = () => <QueryClientProvider client={queryClient}>
         </ARIAProvider>
       </ColorBlindProvider>
     </SoundProvider>
+    </ThemeProvider>
   </QueryClientProvider>;
 export default App;
